+++ libraries/exoplayer_hls/src/test/java/androidx/media3/exoplayer/hls/HlsContentSteeringTrackerTest.java
@@ -25,6 +25,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.clearInvocations;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
@@ -34,6 +35,7 @@ import android.net.Uri;
 import android.os.Looper;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
+import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.Clock;
 import androidx.media3.common.util.SystemClock;
 import androidx.media3.datasource.DataSpec;
@@ -74,16 +76,8 @@ public final class HlsContentSteeringTrackerTest {
 
   @Before
   public void setUp() {
-    HlsRedundantGroup.GroupKey groupKey =
-        new HlsRedundantGroup.GroupKey(new Format.Builder().build(), "id");
-    HlsRedundantGroup group =
-        new HlsRedundantGroup(groupKey, "CDN-A", Uri.parse("https://cdn-a.test"));
-    group.put("CDN-B", Uri.parse("https://cdn-b.test"));
-    group.put("CDN-C", Uri.parse("https://cdn-c.test"));
-
+    setupHlsPlaylistTracker();
     when(mockBandwidthMeter.getBitrateEstimate()).thenReturn(500_000L);
-    when(mockPlaylistTracker.getRedundantGroups(HlsRedundantGroup.VARIANT))
-        .thenReturn(ImmutableList.of(group));
     currentPathwayUpdateCount = new AtomicInteger();
     doAnswer(invocation -> currentPathwayUpdateCount.incrementAndGet())
         .when(mockContentSteeringTrackerCallback)
@@ -179,6 +173,254 @@ public final class HlsContentSteeringTrackerTest {
             /* currentPathwayId= */ eq("CDN-C"),
             /* previousPathwayId= */ eq("CDN-A"),
             /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    assertThat(contentSteeringTracker.isActive()).isTrue();
+  }
+
+  @Test
+  public void start_onSteeringManifestUpdatedWithPathwayClones_notifiesCallback() {
+    String steeringManifest =
+        "{\n"
+            + "  \"VERSION\": 1,\n"
+            + "  \"PATHWAY-PRIORITY\": [\"CDN-A-CLONE\", \"CDN-A\", \"CDN-B\", \"CDN-C\"],\n"
+            + "  \"PATHWAY-CLONES\": [\n"
+            + "    {\n"
+            + "      \"BASE-ID\": \"CDN-A\",\n"
+            + "      \"ID\": \"CDN-A-CLONE\",\n"
+            + "      \"URI-REPLACEMENT\": {\n"
+            + "        \"HOST\": \"backup.com\",\n"
+            + "        \"PARAMS\": {\n"
+            + "          \"param1\": \"newValue1\",\n"
+            + "          \"param2\": \"newValue2\"\n"
+            + "        },\n"
+            + "        \"PER-VARIANT-URIS\": {\n"
+            + "          \"VIDEO-720\": \"https://backup2.com/720.test\"\n"
+            + "        },\n"
+            + "        \"PER-RENDITION-URIS\": {\n"
+            + "          \"AUDIO\": \"https://backup2.com/audio.test\"\n"
+            + "        }\n"
+            + "      }\n"
+            + "    }\n"
+            + "  ]\n"
+            + "}";
+    FakeDataSource fakeDataSource =
+        new FakeDataSource(
+            new FakeDataSet()
+                .newDefaultData()
+                .appendReadData(getBytes(steeringManifest))
+                .endData());
+
+    AtomicReference<TimeoutException> timeoutExceptionRef = new AtomicReference<>();
+    HlsContentSteeringTracker contentSteeringTracker =
+        runContentSteeringTracker(
+            fakeDataSource,
+            /* initialPathwayId= */ "CDN-A",
+            mockContentSteeringTrackerCallback,
+            SystemClock.DEFAULT,
+            /* awaitedCurrentPathwayUpdateCount= */ 2,
+            timeoutExceptionRef);
+
+    assertThat(timeoutExceptionRef.get()).isNull();
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A"),
+            /* previousPathwayId= */ eq(null),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+
+    ImmutableList.Builder<Uri> expectedNewPlaylistUrls = new ImmutableList.Builder<>();
+    ImmutableList.Builder<Uri> expectedBasePlaylistUrls = new ImmutableList.Builder<>();
+    // The 720p variant playlist url and audio rendition playlist url (the first and third in the
+    // list) should be fully overridden by the new urls declared in the PER-VARIANT-URIS and
+    // PER-RENDITION-URIS. The 360p variant playlist url (the second in the list) should have its
+    // host component to be replaced by the HOST, and parameters declared in PARAMS appended as the
+    // query parameters.
+    expectedNewPlaylistUrls.add(
+        Uri.parse("https://backup2.com/720.test"),
+        Uri.parse("https://backup.com/360.test?param1=newValue1&param2=newValue2"),
+        Uri.parse("https://backup2.com/audio.test"));
+    expectedBasePlaylistUrls.add(
+        Uri.parse("https://cdn-a/720.test"),
+        Uri.parse("https://cdn-a/360.test?param1=value1"),
+        Uri.parse("https://cdn-a/audio.test"));
+    verify(mockContentSteeringTrackerCallback)
+        .onNewPathwayAvailable(
+            /* newPathwayId= */ eq("CDN-A-CLONE"),
+            /* basePathwayId= */ eq("CDN-A"),
+            /* newPlaylistUrls= */ eq(expectedNewPlaylistUrls.build()),
+            /* basePlaylistUrls= */ eq(expectedBasePlaylistUrls.build()));
+
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A-CLONE"),
+            /* previousPathwayId= */ eq("CDN-A"),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    assertThat(contentSteeringTracker.isActive()).isTrue();
+  }
+
+  @Test
+  public void start_onSteeringManifestUpdatedWithPerVariantUrisPathwayCloneOnly_notifiesCallback() {
+    String steeringManifest =
+        "{\n"
+            + "  \"VERSION\": 1,\n"
+            + "  \"PATHWAY-PRIORITY\": [\"CDN-A-CLONE\", \"CDN-A\", \"CDN-B\", \"CDN-C\"],\n"
+            + "  \"PATHWAY-CLONES\": [\n"
+            + "    {\n"
+            + "      \"BASE-ID\": \"CDN-A\",\n"
+            + "      \"ID\": \"CDN-A-CLONE\",\n"
+            + "      \"URI-REPLACEMENT\": {\n"
+            + "        \"PER-VARIANT-URIS\": {\n"
+            + "          \"VIDEO-720\": \"https://backup2.com/720.test\"\n"
+            + "        }\n"
+            + "      }\n"
+            + "    }\n"
+            + "  ]\n"
+            + "}";
+    FakeDataSource fakeDataSource =
+        new FakeDataSource(
+            new FakeDataSet()
+                .newDefaultData()
+                .appendReadData(getBytes(steeringManifest))
+                .endData());
+
+    AtomicReference<TimeoutException> timeoutExceptionRef = new AtomicReference<>();
+    HlsContentSteeringTracker contentSteeringTracker =
+        runContentSteeringTracker(
+            fakeDataSource,
+            /* initialPathwayId= */ "CDN-A",
+            mockContentSteeringTrackerCallback,
+            SystemClock.DEFAULT,
+            /* awaitedCurrentPathwayUpdateCount= */ 2,
+            timeoutExceptionRef);
+
+    assertThat(timeoutExceptionRef.get()).isNull();
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A"),
+            /* previousPathwayId= */ eq(null),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+
+    ImmutableList.Builder<Uri> expectedNewPlaylistUrls = new ImmutableList.Builder<>();
+    ImmutableList.Builder<Uri> expectedBasePlaylistUrls = new ImmutableList.Builder<>();
+    // The 720p variant playlist url (the first in the list) should be fully overridden by the new
+    // url declared in the PER-VARIANT-URIS. The 360p variant playlist url and the audio rendition
+    // playlist url (the second and third in the list) should be simply copied from the base urls
+    // as there is no other general URI replacement instruction provided.
+    expectedNewPlaylistUrls.add(
+        Uri.parse("https://backup2.com/720.test"),
+        Uri.parse("https://cdn-a/360.test?param1=value1"),
+        Uri.parse("https://cdn-a/audio.test"));
+    expectedBasePlaylistUrls.add(
+        Uri.parse("https://cdn-a/720.test"),
+        Uri.parse("https://cdn-a/360.test?param1=value1"),
+        Uri.parse("https://cdn-a/audio.test"));
+    verify(mockContentSteeringTrackerCallback)
+        .onNewPathwayAvailable(
+            /* newPathwayId= */ eq("CDN-A-CLONE"),
+            /* basePathwayId= */ eq("CDN-A"),
+            /* newPlaylistUrls= */ eq(expectedNewPlaylistUrls.build()),
+            /* basePlaylistUrls= */ eq(expectedBasePlaylistUrls.build()));
+
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A-CLONE"),
+            /* previousPathwayId= */ eq("CDN-A"),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    assertThat(contentSteeringTracker.isActive()).isTrue();
+  }
+
+  @Test
+  public void start_onSteeringManifestUpdatedWithPathwayCloneWithUnknownBaseId_skipsClone() {
+    String steeringManifest =
+        "{\n"
+            + "  \"VERSION\": 1,\n"
+            + "  \"PATHWAY-PRIORITY\": [\"CDN-D-CLONE\", \"CDN-A\", \"CDN-B\", \"CDN-C\"],\n"
+            + "  \"PATHWAY-CLONES\": [\n"
+            + "    {\n"
+            + "      \"BASE-ID\": \"CDN-D\",\n"
+            + "      \"ID\": \"CDN-D-CLONE\",\n"
+            + "      \"URI-REPLACEMENT\": {\n"
+            + "        \"PER-VARIANT-URIS\": {\n"
+            + "          \"VIDEO-720\": \"https://backup2.com/720.test\"\n"
+            + "        }\n"
+            + "      }\n"
+            + "    }\n"
+            + "  ]\n"
+            + "}";
+    FakeDataSource fakeDataSource =
+        new FakeDataSource(
+            new FakeDataSet()
+                .newDefaultData()
+                .appendReadData(getBytes(steeringManifest))
+                .endData());
+
+    AtomicReference<TimeoutException> timeoutExceptionRef = new AtomicReference<>();
+    HlsContentSteeringTracker contentSteeringTracker =
+        runContentSteeringTracker(
+            fakeDataSource,
+            /* initialPathwayId= */ "CDN-A",
+            mockContentSteeringTrackerCallback,
+            SystemClock.DEFAULT,
+            /* awaitedCurrentPathwayUpdateCount= */ 1,
+            timeoutExceptionRef);
+
+    assertThat(timeoutExceptionRef.get()).isNull();
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A"),
+            /* previousPathwayId= */ eq(null),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    verify(mockContentSteeringTrackerCallback, never())
+        .onNewPathwayAvailable(any(), any(), any(), any());
+    assertThat(contentSteeringTracker.isActive()).isTrue();
+  }
+
+  @Test
+  public void start_onSteeringManifestUpdatedWithPathwayCloneWithExistingId_skipsClone() {
+    String steeringManifest =
+        "{\n"
+            + "  \"VERSION\": 1,\n"
+            + "  \"PATHWAY-PRIORITY\": [\"CDN-B\", \"CDN-A\", \"CDN-C\"],\n"
+            + "  \"PATHWAY-CLONES\": [\n"
+            + "    {\n"
+            + "      \"BASE-ID\": \"CDN-A\",\n"
+            + "      \"ID\": \"CDN-B\",\n"
+            + "      \"URI-REPLACEMENT\": {\n"
+            + "        \"PER-VARIANT-URIS\": {\n"
+            + "          \"VIDEO-720\": \"https://backup2.com/720.test\"\n"
+            + "        }\n"
+            + "      }\n"
+            + "    }\n"
+            + "  ]\n"
+            + "}";
+    FakeDataSource fakeDataSource =
+        new FakeDataSource(
+            new FakeDataSet()
+                .newDefaultData()
+                .appendReadData(getBytes(steeringManifest))
+                .endData());
+
+    AtomicReference<TimeoutException> timeoutExceptionRef = new AtomicReference<>();
+    HlsContentSteeringTracker contentSteeringTracker =
+        runContentSteeringTracker(
+            fakeDataSource,
+            /* initialPathwayId= */ "CDN-A",
+            mockContentSteeringTrackerCallback,
+            SystemClock.DEFAULT,
+            /* awaitedCurrentPathwayUpdateCount= */ 2,
+            timeoutExceptionRef);
+
+    assertThat(timeoutExceptionRef.get()).isNull();
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-A"),
+            /* previousPathwayId= */ eq(null),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    verify(mockContentSteeringTrackerCallback)
+        .onCurrentPathwayUpdated(
+            /* currentPathwayId= */ eq("CDN-B"),
+            /* previousPathwayId= */ eq("CDN-A"),
+            /* previousPathwayExcludeDurationMs= */ eq(C.TIME_UNSET));
+    verify(mockContentSteeringTrackerCallback, never())
+        .onNewPathwayAvailable(any(), any(), any(), any());
     assertThat(contentSteeringTracker.isActive()).isTrue();
   }
 
@@ -449,6 +691,43 @@ public final class HlsContentSteeringTrackerTest {
     // Exclude CDN-A
     assertThat(contentSteeringTracker.excludeCurrentPathway(10_000L)).isFalse();
     verifyNoInteractions(mockContentSteeringTrackerCallback);
+  }
+
+  private void setupHlsPlaylistTracker() {
+    HlsRedundantGroup.GroupKey variantGroupKey1 =
+        new HlsRedundantGroup.GroupKey(
+            new Format.Builder().setSampleMimeType(MimeTypes.VIDEO_H264).build(),
+            /* stableId= */ "VIDEO-720");
+    HlsRedundantGroup.GroupKey variantGroupKey2 =
+        new HlsRedundantGroup.GroupKey(
+            new Format.Builder().setSampleMimeType(MimeTypes.VIDEO_H264).build(),
+            /* stableId= */ "VIDEO-360");
+    HlsRedundantGroup.GroupKey audioGroupKey =
+        new HlsRedundantGroup.GroupKey(
+            new Format.Builder().setSampleMimeType(MimeTypes.AUDIO_AAC).build(),
+            /* stableId= */ "AUDIO");
+    HlsRedundantGroup variantRedundantGroup1 =
+        new HlsRedundantGroup(variantGroupKey1, "CDN-A", Uri.parse("https://cdn-a/720.test"));
+    variantRedundantGroup1.put("CDN-B", Uri.parse("https://cdn-b/720.test"));
+    variantRedundantGroup1.put("CDN-C", Uri.parse("https://cdn-c/720.test"));
+    HlsRedundantGroup variantRedundantGroup2 =
+        new HlsRedundantGroup(
+            variantGroupKey2, "CDN-A", Uri.parse("https://cdn-a/360.test?param1=value1"));
+    variantRedundantGroup2.put("CDN-B", Uri.parse("https://cdn-b/360.test"));
+    variantRedundantGroup2.put("CDN-C", Uri.parse("https://cdn-c/360.test"));
+    HlsRedundantGroup audioRedundantGroup =
+        new HlsRedundantGroup(audioGroupKey, "CDN-A", Uri.parse("https://cdn-a/audio.test"));
+    audioRedundantGroup.put("CDN-B", Uri.parse("https://cdn-b/audio.test"));
+    audioRedundantGroup.put("CDN-C", Uri.parse("https://cdn-c/audio.test"));
+
+    when(mockPlaylistTracker.getRedundantGroups(HlsRedundantGroup.VARIANT))
+        .thenReturn(ImmutableList.of(variantRedundantGroup1, variantRedundantGroup2));
+    when(mockPlaylistTracker.getRedundantGroups(HlsRedundantGroup.AUDIO_RENDITION))
+        .thenReturn(ImmutableList.of(audioRedundantGroup));
+    when(mockPlaylistTracker.getRedundantGroups(HlsRedundantGroup.VIDEO_RENDITION))
+        .thenReturn(ImmutableList.of());
+    when(mockPlaylistTracker.getRedundantGroups(HlsRedundantGroup.SUBTITLE_RENDITION))
+        .thenReturn(ImmutableList.of());
   }
 
   private HlsContentSteeringTracker runContentSteeringTracker(